refactor(avatar): one Avatar primitive + one pure initials util + composing wrappers (R3 Phase 1 core)
The canonical-component discipline (Zach's canary). The Avatar primitive already
existed + was adopted ~25 sites; this collapses the remaining divergence so
identity rendering has exactly one home and zero random reimplementations.

- src/lib/initials.ts — the ONE initials algorithm (pure, framework-free so
  server code imports it too). Collapses 4 hand-rolled copies: Avatar.initialsFor,
  strokePlayConfig.initialsOf, news.ts initialsOf, CrewRoster PlaceholderAvatar
  inline, + a 5th found mid-refactor (MemberEditor inline).
- Avatar.tsx is the one home for identity rendering: it owns the primitive AND
  the named wrappers (InvitedAvatar, PlaceholderAvatar) that compose it — moved
  here from CrewRoster (re-exported there for back-compat).
- Folded 3 InvitedAvatar copies (CrewRoster + CrewEmailPanel + MemberEditor
  inline) into one wrapper (size + ringColor props). 2 PlaceholderAvatar copies
  (CrewRoster + MemberEditor inline) -> <Avatar muted>.
- strokePlayConfig.initialsOf is now a deprecated alias of initialsFor (one
  algorithm); removed with Participant.initials in the chip-migration follow-up.

NOT in this PR: the scorecard identity chips (StandardGrid/ScoreEntryView/
FinalStandings/MatchEntryView) still render bespoke {p.initials} circles — they
need a new Avatar chip variant (faint-bg/colored-fg) + touch the scorecard hot
path, so per the spec they're a tight follow-up PR (still R3), with their own
scorecard verification.

Verified: tsc + eslint clean; initials + news unit tests pass; E2E critical-path
green; crew tab visual check (avatars + invited badge + legend render identically).
