+++ src/lib/strokePlayConfig.ts
@@ -1,4 +1,5 @@
 import type { ScoreUnit } from "@/components/games/types";
+import { initialsFor } from "@/lib/initials";
 
 /**
  * Client-side stroke-play scorecard config, shared by the DB-backed trip game
@@ -89,8 +90,9 @@ export function strokeIndexOf(units: ScoreUnit[]): number[] {
 // team colors per STYLE_GUIDE §7).
 export const PLAYER_COLORS = ["#2dd4bf", "#60a5fa", "#f59e0b", "#a855f7"];
 
-/** 1–2 char initials from a free-text name (falls back to "?"). */
-export function initialsOf(name: string): string {
-  const parts = name.trim().split(/\s+/);
-  return ((parts[0]?.[0] ?? "") + (parts[1]?.[0] ?? "")).toUpperCase() || "?";
-}
+/**
+ * @deprecated Import `initialsFor` from `@/lib/initials` directly. This alias
+ * remains only while `Participant.initials` is precomputed for the scorecard
+ * chips; it goes away when those route through the Avatar primitive.
+ */
+export const initialsOf = initialsFor;
